+++ extensions/integration-tests/src/onboarding/developers.ts
@@ -11,6 +11,7 @@ export const developers: string[] = [
 	'alanrenmsft',
 	'anjalia',
 	'anthonydresser',
+	'caohai',
 	'Charles-Gagnon',
 	'cssuh',
 	'ktech99',