+++ src/sql/platform/theme/common/styler.ts
@@ -8,7 +8,7 @@ import * as colors from './colors';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
 import * as cr from 'vs/platform/theme/common/colorRegistry';
 import * as sqlcr from 'sql/platform/theme/common/colorRegistry';
-import { attachStyler, IColorMapping, IStyleOverrides } from 'vs/platform/theme/common/styler';
+import { attachStyler, defaultListStyles, IColorMapping, IStyleOverrides } from 'vs/platform/theme/common/styler';
 import { IDisposable } from 'vs/base/common/lifecycle';
 import { IThemable } from 'vs/base/common/styler';
 
@@ -325,4 +325,27 @@ export const defaultInfoButtonStyles: IInfoButtonStyleOverrides = {
 
 export function attachInfoButtonStyler(widget: IThemable, themeService: IThemeService, style?: IInfoButtonStyleOverrides): IDisposable {
 	return attachStyler(themeService, { ...defaultInfoButtonStyles, ...style }, widget);
+}
+
+export function attachTableFilterStyler(widget: IThemable, themeService: IThemeService): IDisposable {
+	return attachStyler(themeService, {
+		inputBackground: cr.inputBackground,
+		inputForeground: cr.inputForeground,
+		inputBorder: cr.inputBorder,
+		buttonForeground: cr.buttonForeground,
+		buttonBackground: cr.buttonBackground,
+		buttonHoverBackground: cr.buttonHoverBackground,
+		buttonSecondaryForeground: cr.buttonSecondaryForeground,
+		buttonSecondaryBackground: cr.buttonSecondaryBackground,
+		buttonSecondaryHoverBackground: cr.buttonSecondaryHoverBackground,
+		buttonBorder: cr.buttonBorder,
+		buttonSecondaryBorder: cr.buttonSecondaryBorder,
+		buttonDisabledBorder: cr.buttonDisabledBorder,
+		buttonDisabledBackground: cr.buttonDisabledBackground,
+		buttonDisabledForeground: cr.buttonDisabledForeground,
+		badgeBackground: cr.badgeBackground,
+		badgeForeground: cr.badgeForeground,
+		badgeBorder: cr.contrastBorder,
+		...defaultListStyles,
+	}, widget);
 }