+++ src/sql/workbench/services/connection/browser/connectionBrowseTab.ts
@@ -12,6 +12,8 @@ import { ConnectionProfile } from 'sql/platform/connection/common/connectionProf
 import { ConnectionProfileGroup } from 'sql/platform/connection/common/connectionProfileGroup';
 import { attachInputBoxStyler } from 'sql/platform/theme/common/styler';
 import { ITreeItem } from 'sql/workbench/common/views';
+import { CONNECTIONS_SORT_BY_CONFIG_KEY } from 'sql/platform/connection/common/connectionConfig';
+import { ConnectionSource } from 'sql/workbench/services/connection/browser/connectionDialogWidget';
 import { IConnectionTreeDescriptor, IConnectionTreeService } from 'sql/workbench/services/connection/common/connectionTreeService';
 import { AsyncRecentConnectionTreeDataSource } from 'sql/workbench/services/objectExplorer/browser/asyncRecentConnectionTreeDataSource';
 import { ServerTreeElement } from 'sql/workbench/services/objectExplorer/browser/asyncServerTree';
@@ -20,12 +22,13 @@ import { ServerTreeRenderer } from 'sql/workbench/services/objectExplorer/browse
 import { TreeUpdateUtils } from 'sql/workbench/services/objectExplorer/browser/treeUpdateUtils';
 import { TreeNode } from 'sql/workbench/services/objectExplorer/common/treeNode';
 import * as DOM from 'vs/base/browser/dom';
+import { ActionBar } from 'vs/base/browser/ui/actionbar/actionbar';
 import { status } from 'vs/base/browser/ui/aria/aria';
 import { IIdentityProvider, IListVirtualDelegate } from 'vs/base/browser/ui/list/list';
 import { IListAccessibilityProvider } from 'vs/base/browser/ui/list/listWidget';
 import { ProgressBar } from 'vs/base/browser/ui/progressbar/progressbar';
 import { IAsyncDataSource, ITreeNode, ITreeRenderer } from 'vs/base/browser/ui/tree/tree';
-import { IAction } from 'vs/base/common/actions';
+import { IAction, IActionViewItemProvider } from 'vs/base/common/actions';
 import { debounce } from 'vs/base/common/decorators';
 import { Emitter, Event } from 'vs/base/common/event';
 import { Iterable } from 'vs/base/common/iterator';
@@ -35,8 +38,8 @@ import { isString } from 'vs/base/common/types';
 import { URI } from 'vs/base/common/uri';
 import 'vs/css!./media/connectionBrowseTab';
 import { localize } from 'vs/nls';
-import { createAndFillInContextMenuActions } from 'vs/platform/actions/browser/menuEntryActionViewItem';
-import { IMenuService, MenuId } from 'vs/platform/actions/common/actions';
+import { createAndFillInContextMenuActions, MenuEntryActionViewItem } from 'vs/platform/actions/browser/menuEntryActionViewItem';
+import { IMenuService, MenuId, MenuItemAction } from 'vs/platform/actions/common/actions';
 import { ICommandService } from 'vs/platform/commands/common/commands';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { IContextKey, IContextKeyService, RawContextKey } from 'vs/platform/contextkey/common/contextkey';
@@ -56,7 +59,7 @@ import { ITreeItemLabel, ITreeViewDataProvider, TreeItemCollapsibleState, TreeVi
 export type TreeElement = ConnectionDialogTreeProviderElement | ITreeItemFromProvider | SavedConnectionNode | ServerTreeElement;
 
 export class ConnectionBrowseTab implements IPanelTab {
-	public readonly title = localize('connectionDialog.browser', "Browse (Preview)");
+	public readonly title = localize('connectionDialog.browser', "Browse");
 	public readonly identifier = 'connectionBrowse';
 	public readonly view = this.instantiationService.createInstance(ConnectionBrowserView);
 	constructor(@IInstantiationService private readonly instantiationService: IInstantiationService) { }
@@ -64,7 +67,8 @@ export class ConnectionBrowseTab implements IPanelTab {
 
 export interface SelectedConnectionChangedEventArgs {
 	connectionProfile: IConnectionProfile,
-	connect: boolean
+	connect: boolean,
+	source: ConnectionSource
 }
 
 export class ConnectionBrowserView extends Disposable implements IPanelView {
@@ -73,8 +77,8 @@ export class ConnectionBrowserView extends Disposable implements IPanelView {
 	private treeContainer: HTMLElement | undefined;
 	private model: TreeModel | undefined;
 	private treeLabels: ResourceLabels | undefined;
+	private treeMenus: ConnectionBrowseTreeMenuProvider | undefined;
 	private treeDataSource: DataSource | undefined;
-	private readonly contextKey = new ContextKey(this.contextKeyService);
 	private filterProgressBar: ProgressBar | undefined;
 
 	public onDidChangeVisibility = Event.None;
@@ -88,11 +92,10 @@ export class ConnectionBrowserView extends Disposable implements IPanelView {
 		@IContextViewService private readonly contextViewService: IContextViewService,
 		@IThemeService private readonly themeService: IThemeService,
 		@ICommandService private readonly commandService: ICommandService,
-		@IMenuService private readonly menuService: IMenuService,
-		@IContextKeyService private readonly contextKeyService: IContextKeyService,
 		@IContextMenuService private readonly contextMenuService: IContextMenuService,
 		@IConnectionManagementService private readonly connectionManagementService: IConnectionManagementService,
-		@ICapabilitiesService private readonly capabilitiesService: ICapabilitiesService
+		@ICapabilitiesService private readonly capabilitiesService: ICapabilitiesService,
+		@IConfigurationService private readonly configurationService: IConfigurationService,
 	) {
 		super();
 		this.connectionTreeService.setView(this);
@@ -150,9 +153,16 @@ export class ConnectionBrowserView extends Disposable implements IPanelView {
 	renderTree(container: HTMLElement): void {
 		this.treeContainer = container.appendChild(DOM.$('div'));
 		this.treeLabels = this._register(this.instantiationService.createInstance(ResourceLabels, this));
+		this.treeMenus = this.instantiationService.createInstance(ConnectionBrowseTreeMenuProvider);
+		const actionViewItemProvider = (action: IAction) => {
+			if (action instanceof MenuItemAction) {
+				return this.instantiationService.createInstance(MenuEntryActionViewItem, action);
+			}
+			return undefined;
+		};
 		const renderers: ITreeRenderer<TreeElement, any, any>[] = [
-			new ProviderElementRenderer(),
-			this.instantiationService.createInstance(TreeItemRenderer, this.treeLabels),
+			new ProviderElementRenderer(this.treeMenus, actionViewItemProvider),
+			this.instantiationService.createInstance(TreeItemRenderer, this.treeMenus, this.treeLabels, actionViewItemProvider),
 			this.instantiationService.createInstance(ConnectionProfileRenderer, true),
 			this.instantiationService.createInstance(ConnectionProfileGroupRenderer),
 			this.instantiationService.createInstance(TreeNodeRenderer),
@@ -176,26 +186,12 @@ export class ConnectionBrowserView extends Disposable implements IPanelView {
 				accessibilityProvider: new ListAccessibilityProvider()
 			}) as WorkbenchAsyncDataTree<TreeModel, TreeElement>);
 		this._register(this.tree.onContextMenu(e => {
-			let context: ITreeItem | ConnectionDialogTreeProviderElement | undefined;
-			let actionContext: TreeViewItemHandleArg | ConnectionDialogTreeProviderElement | undefined;
-			if (instanceOfITreeItemFromProvider(e.element)) {
-				context = e.element.element;
-				actionContext = <TreeViewItemHandleArg>{ $treeViewId: e.element.treeId, $treeItemHandle: context.handle, $treeItem: context };
-			} else if (e.element instanceof ConnectionDialogTreeProviderElement) {
-				context = e.element;
-				actionContext = e.element;
-			}
-			if (context) {
-				this.contextKey.set(context);
-				const menu = this.menuService.createMenu(MenuId.ConnectionDialogBrowseTreeContext, this.contextKeyService);
-				const primary: IAction[] = [];
-				const secondary: IAction[] = [];
-				const result = { primary, secondary };
-				createAndFillInContextMenuActions(menu, { shouldForwardArgs: true }, result);
-
+			const actionContext = this.treeMenus.getActionContext(e.element);
+			if (actionContext) {
+				const actions = this.treeMenus.getActions(e.element);
 				this.contextMenuService.showContextMenu({
 					getAnchor: () => e.anchor,
-					getActions: () => result.primary,
+					getActions: () => actions,
 					getActionsContext: () => actionContext
 				});
 			}
@@ -232,6 +228,13 @@ export class ConnectionBrowserView extends Disposable implements IPanelView {
 		this._register(this.themeService.onDidColorThemeChange(async () => {
 			await this.refresh();
 		}));
+
+		this._register(this.configurationService.onDidChangeConfiguration(e => {
+			if (e.affectsConfiguration(CONNECTIONS_SORT_BY_CONFIG_KEY)) {
+				this.updateSavedConnectionsNode();
+			}
+		}));
+
 	}
 
 	private handleTreeElementSelection(selectedNode: TreeElement, connect: boolean): void {
@@ -246,14 +249,16 @@ export class ConnectionBrowserView extends Disposable implements IPanelView {
 					this._onSelectedConnectionChanged.fire(
 						{
 							connectionProfile: selectedNode.element.payload,
-							connect: connect
+							connect: connect,
+							source: 'azure'
 						});
 				}
 			}
 		} else if (selectedNode instanceof ConnectionProfile) {
 			this._onSelectedConnectionChanged.fire({
 				connectionProfile: selectedNode,
-				connect: connect
+				connect: connect,
+				source: 'savedconnections'
 			});
 		}
 	}
@@ -278,7 +283,7 @@ export class ConnectionBrowserView extends Disposable implements IPanelView {
 	}
 
 	focus(): void {
-		this.tree.domFocus();
+		this.filterInput.focus();
 	}
 }
 
@@ -334,6 +339,7 @@ class ListDelegate implements IListVirtualDelegate<TreeElement> {
 interface TreeElementTemplate {
 	readonly icon: HTMLElement;
 	readonly name: HTMLElement;
+	readonly actionBar?: ActionBar;
 }
 
 abstract class BaseTreeItemRender<T> implements ITreeRenderer<T, void, TreeElementTemplate> {
@@ -341,25 +347,47 @@ abstract class BaseTreeItemRender<T> implements ITreeRenderer<T, void, TreeEleme
 	protected abstract getText(element: T): string;
 	protected abstract getIconClass(element: T): string;
 
+	constructor(private menus?: ConnectionBrowseTreeMenuProvider,
+		private actionViewItemProvider?: IActionViewItemProvider
+	) {
+	}
+
 	renderTemplate(container: HTMLElement): TreeElementTemplate {
 		const root = DOM.append(container, DOM.$('div.browse-tree-element'));
 		const icon = DOM.append(root, DOM.$('.icon'));
 		const name = DOM.append(root, DOM.$('.name'));
-		return { name, icon };
+		let actionBar: ActionBar | undefined;
+		if (this.menus) {
+			const actionsContainer = DOM.append(root, DOM.$('.actions'));
+			actionBar = new ActionBar(actionsContainer, {
+				actionViewItemProvider: this.actionViewItemProvider
+			});
+		}
+		return { name, icon, actionBar };
 	}
 
 	renderElement(element: ITreeNode<T, void>, index: number, templateData: TreeElementTemplate, height: number): void {
 		templateData.name.innerText = this.getText(element.element);
 		templateData.icon.classList.add('codicon', this.getIconClass(element.element));
+		if (this.menus) {
+			templateData.actionBar.clear();
+			templateData.actionBar.context = this.menus.getActionContext(element.element);
+			templateData.actionBar.push(this.menus.getActions(element.element), { icon: true, label: false });
+		}
 	}
 
 	disposeTemplate(templateData: TreeElementTemplate): void {
+		templateData.actionBar?.dispose();
 	}
 }
 
 class ProviderElementRenderer extends BaseTreeItemRender<ConnectionDialogTreeProviderElement> {
 	public static readonly TEMPLATE_ID = 'ProviderElementTemplate';
 	public readonly templateId = ProviderElementRenderer.TEMPLATE_ID;
+
+	constructor(menus: ConnectionBrowseTreeMenuProvider, actionViewItemProvider: IActionViewItemProvider) {
+		super(menus, actionViewItemProvider);
+	}
 
 	getText(element: ConnectionDialogTreeProviderElement): string {
 		return element.name;
@@ -538,21 +566,66 @@ interface ITreeExplorerTemplateData {
 	container: HTMLElement;
 	resourceLabel: IResourceLabel;
 	icon: HTMLElement;
-	// actionBar: ActionBar;
+	actionBar: ActionBar;
+}
+
+class ConnectionBrowseTreeMenuProvider {
+	constructor(
+		@IMenuService private readonly menuService: IMenuService,
+		@IContextKeyService private readonly contextKeyService: IContextKeyService
+	) {
+	}
+
+	public getActionContext(element: any): TreeViewItemHandleArg | ConnectionDialogTreeProviderElement | undefined {
+		let actionContext: TreeViewItemHandleArg | ConnectionDialogTreeProviderElement | undefined;
+		if (instanceOfITreeItemFromProvider(element)) {
+			actionContext = <TreeViewItemHandleArg>{ $treeViewId: element.treeId, $treeItemHandle: element.element.handle, $treeItem: element.element };
+		} else if (element instanceof ConnectionDialogTreeProviderElement) {
+			actionContext = element;
+		}
+		return actionContext;
+	}
+
+	public getActions(element: any): IAction[] {
+		if (!this.contextKeyService) {
+			return [];
+		}
+
+		let context: ITreeItem | ConnectionDialogTreeProviderElement | undefined;
+		if (instanceOfITreeItemFromProvider(element)) {
+			context = element.element;
+		} else if (element instanceof ConnectionDialogTreeProviderElement) {
+			context = element;
+		}
+
+		if (!context) {
+			return [];
+		}
+
+		const contextKeyService = this.contextKeyService.createScoped();
+		const contextKey = new ContextKey(contextKeyService);
+		contextKey.set(context);
+		const menu = this.menuService.createMenu(MenuId.ConnectionDialogBrowseTreeContext, contextKeyService);
+		const primary: IAction[] = [];
+		const secondary: IAction[] = [];
+		const result = { primary, secondary };
+		createAndFillInContextMenuActions(menu, { shouldForwardArgs: true }, result);
+
+		menu.dispose();
+		contextKeyService.dispose();
+
+		return result.primary;
+	}
 }
 
 class TreeItemRenderer extends Disposable implements ITreeRenderer<ITreeItemFromProvider, void, ITreeExplorerTemplateData> {
 	static readonly ITEM_HEIGHT = 22;
 	static readonly TREE_TEMPLATE_ID = 'treeExplorer';
 
-	// private _actionRunner: MultipleSelectionActionRunner | undefined;
-
 	constructor(
-		// private treeViewId: string,
-		// private menus: TreeMenus,
+		private menus: ConnectionBrowseTreeMenuProvider,
 		private labels: ResourceLabels,
-		// private actionViewItemProvider: IActionViewItemProvider,
-		// private aligner: Aligner,
+		private actionViewItemProvider: IActionViewItemProvider,
 		@IThemeService private readonly themeService: IThemeService,
 		@IConfigurationService private readonly configurationService: IConfigurationService,
 		@ILabelService private readonly labelService: ILabelService
@@ -564,22 +637,18 @@ class TreeItemRenderer extends Disposable implements ITreeRenderer<ITreeItemFrom
 		return TreeItemRenderer.TREE_TEMPLATE_ID;
 	}
 
-	// set actionRunner(actionRunner: MultipleSelectionActionRunner) {
-	// 	this._actionRunner = actionRunner;
-	// }
-
 	renderTemplate(container: HTMLElement): ITreeExplorerTemplateData {
 		container.classList.add('custom-view-tree-node-item');
 
 		const icon = DOM.append(container, DOM.$('.custom-view-tree-node-item-icon'));
 
 		const resourceLabel = this.labels.create(container, { supportHighlights: true });
-		// const actionsContainer = DOM.append(resourceLabel.element, DOM.$('.actions'));
-		// const actionBar = new ActionBar(actionsContainer, {
-		// 	actionViewItemProvider: this.actionViewItemProvider
-		// });
+		const actionsContainer = DOM.append(resourceLabel.element, DOM.$('.actions'));
+		const actionBar = new ActionBar(actionsContainer, {
+			actionViewItemProvider: this.actionViewItemProvider
+		});
 
-		return { resourceLabel, icon, container, elementDisposable: Disposable.None };
+		return { resourceLabel, icon, container, elementDisposable: Disposable.None, actionBar };
 	}
 
 	renderElement(element: ITreeNode<ITreeItemFromProvider, void>, index: number, templateData: ITreeExplorerTemplateData): void {
@@ -595,7 +664,7 @@ class TreeItemRenderer extends Disposable implements ITreeRenderer<ITreeItemFrom
 		const sqlIcon = node.sqlIcon;
 
 		// reset
-		// templateData.actionBar.clear();
+		templateData.actionBar.clear();
 
 		if (resource || this.isFileKindThemeIcon(node.themeIcon)) {
 			const fileDecorations = this.configurationService.getValue<{ colors: boolean, badges: boolean }>('explorer.decorations');
@@ -622,17 +691,8 @@ class TreeItemRenderer extends Disposable implements ITreeRenderer<ITreeItemFrom
 			templateData.icon.style.backgroundImage = '';
 		}
 
-		// templateData.actionBar.context = <TreeViewItemHandleArg>{ $treeViewId: this.treeViewId, $treeItemHandle: node.handle };
-		// templateData.actionBar.push(this.menus.getResourceActions(node), { icon: true, label: false });
-		// if (this._actionRunner) {
-		// 	templateData.actionBar.actionRunner = this._actionRunner;
-		// }
-		this.setAlignment(templateData.container, node);
-		templateData.elementDisposable = (this.themeService.onDidFileIconThemeChange(() => this.setAlignment(templateData.container, node)));
-	}
-
-	private setAlignment(container: HTMLElement, treeItem: ITreeItem) {
-		// DOM.toggleClass(container.parentElement!, 'align-icon-with-twisty', this.aligner.alignIconWithTwisty(treeItem));
+		templateData.actionBar.context = this.menus.getActionContext(element.element);
+		templateData.actionBar.push(this.menus.getActions(element.element), { icon: true, label: false });
 	}
 
 	private isFileKindThemeIcon(icon: ThemeIcon | undefined): boolean {
@@ -660,8 +720,8 @@ class TreeItemRenderer extends Disposable implements ITreeRenderer<ITreeItemFrom
 	}
 
 	disposeTemplate(templateData: ITreeExplorerTemplateData): void {
-		// templateData.resourceLabel.dispose();
-		// templateData.actionBar.dispose();
+		templateData.resourceLabel.dispose();
+		templateData.actionBar.dispose();
 		templateData.elementDisposable.dispose();
 	}
 }